requirements.txt
fastapi
uvicorn
streamlit
groq
python-dotenv
httpx
pydantic
